+++ packages/coding-standards/tests/Unit/CsFixer/OrmJoinColumnRequireNullableFixer/correct/many_to_one_missing_join_column.php
@@ -1,0 +1,9 @@
+<?php
+
+class Bar
+{
+    /**
+     * @var int
+     */
+    private $number;
+}