+++ packages/coding-standards/tests/Unit/CsFixer/OrmJoinColumnRequireNullableFixer/fixed/many_to_one_missing_nullable_param.php
@@ -1,0 +1,11 @@
+<?php
+
+class Bar
+{
+    /**
+     * @var \StdObject
+     * @ORM\ManyToOne(targetEntity="StdObject")
+     * @ORM\JoinColumn(nullable=false, name="std_id", referencedColumnName="id")
+     */
+    private $foo3;
+}