+++ packages/coding-standards/tests/Unit/CsFixer/OrmJoinColumnRequireNullableFixer/wrong/many_to_one_missing_join_column.php
@@ -1,0 +1,10 @@
+<?php
+
+class Bar
+{
+    /**
+     * @var \StdObject
+     * @ORM\ManyToOne(targetEntity="StdObject")
+     */
+    private $foo2;
+}