+++ packages/coding-standards/tests/Unit/CsFixer/OrmJoinColumnRequireNullableFixer/wrong/one_to_one_multiline_missing_nullable_param.php
@@ -1,0 +1,14 @@
+<?php
+
+class Bar
+{
+    /**
+     * @var \StdObject
+     * @ORM\OneToOne(targetEntity="StdObject")
+     * @ORM\JoinColumn(
+     *     name="std_id",
+     *     referencedColumnName="id"
+     * )
+     */
+    private $foo8;
+}