+++ packages/coding-standards/tests/Unit/Sniffs/ForbiddenSuperGlobalSniff/ForbiddenSuperGlobalSniffTest.php
@@ -1,0 +1,21 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Tests\CodingStandards\Sniffs\ForbiddenSuperGlobalSniff;
+
+use Symplify\EasyCodingStandardTester\Testing\AbstractCheckerTestCase;
+
+final class ForbiddenSuperGlobalSniffTest extends AbstractCheckerTestCase
+{
+    public function testWrong(): void
+    {
+        $this->doTestWrongFile(__DIR__ . '/wrong/env.php.inc');
+        $this->doTestWrongFile(__DIR__ . '/wrong/post.php.inc');
+    }
+
+    protected function provideConfig(): string
+    {
+        return __DIR__ . '/config.yml';
+    }
+}